discord_integration.py
# discord_integration.py

import discord
from discord.ext import commands

bot = commands.Bot(command_prefix='!')

# Example function to fetch server information from Discord API
def get_server_info(guild_id):
    # Initialize bot (ensure it's ready before fetching guild)
    @bot.event
    async def on_ready():
        print(f'Bot logged in as {bot.user}')
    
    # Fetch guild information
    guild = bot.get_guild(guild_id)
    if guild:
        member_count = guild.member_count
        channels = len(guild.channels)
        roles = len(guild.roles)
        server_info = {
            'guild_name': guild.name,
            'member_count': member_count,
            'channel_count': channels,
            'role_count': roles
        }
        return server_info
    else:
        return None

# Run the bot in standalone mode (for testing)
if __name__ == '__main__':
    bot.run('MTI1MzQ5Njc2ODEzODcwNzA0NQ.GAFm4m.WBgYocuGHAz6g0J7aAyGW89_lz5PFWtn5nPymQ')  # Replace with your actual bot token
